ml-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
numpy==1.24.3
ultralytics==8.0.206
